requirements.txt
ebooklib
beautifulsoup4
fastapi
uvicorn
python-multipart
lameenc
mutagen
typer
rich
pocket-tts  # For local TTS (no API limits)
google-genai  # For Gemini text processing (clean, summarize, speed read)
python-dotenv  # For loading .env file
numpy  # For audio processing
